# atoma-daemon/docs/openapi.yml
openapi: 3.1.0
info:
  title: atoma-daemon
  description: ''
  license:
    name: Apache-2.0
  version: 0.1.0
servers:
- url: http://localhost:8080
  description: Local development server
paths:
  /almost_filled_stacks/{fraction}:
    get:
      tags:
      - almost_filled_stacks
      summary: Retrieves all stacks that are filled above a specified fraction threshold for all registered nodes.
      operationId: almost_filled_stacks_fraction_get
      parameters:
      - name: fraction
        in: path
        description: The fraction threshold (0.0 to 100.0) to filter stacks
        required: true
        schema:
          type: number
          format: double
      responses:
        '200':
          description: Stack objects that are filled above the specified fraction threshold
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stack'
        '500':
          description: Internal server error
  /almost_filled_stacks/{id}/{fraction}:
    get:
      tags:
      - almost_filled_stacks
      summary: Retrieves all stacks that are filled above a specified fraction threshold for a specific node.
      description: |-
        # Arguments
        * `daemon_state` - The shared state containing the state manager
        * `node_small_id` - The small ID of the node whose stacks should be retrieved
        * `fraction` - The fraction threshold (0.0 to 100.0) to filter stacks

        # Returns
        * `Result<Json<Vec<Stack>>>` - A JSON response containing a list of stacks
          - `Ok(Json<Vec<Stack>>)` - Successfully retrieved stacks
          - `Err(StatusCode::INTERNAL_SERVER_ERROR)` - Failed to retrieve stacks from state manager

        # Example Response
        Returns a JSON array of Stack objects for the specified node that are filled above the fraction threshold
      operationId: almost_filled_stacks_get
      parameters:
      - name: id
        in: path
        description: The small ID of the node whose stacks should be retrieved
        required: true
        schema:
          type: integer
          format: int64
      - name: fraction
        in: path
        description: The fraction threshold (0.0 to 100.0) to filter stacks
        required: true
        schema:
          type: number
          format: double
      responses:
        '200':
          description: List of node stacks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stack'
        '500':
          description: Internal server error
  /attestation_disputes/against:
    get:
      tags:
      - attestation_disputes
      summary: Retrieves all attestation disputes against the currently registered nodes.
      operationId: attestation_disputes_against_list
      responses:
        '200':
          description: List of all against attestation disputes where the registered nodes are the defendants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StackAttestationDispute'
        '500':
          description: Internal server error
  /attestation_disputes/against/{id}:
    get:
      tags:
      - attestation_disputes
      summary: Retrieves all attestation disputes against a specific node.
      operationId: attestation_disputes_against_get
      parameters:
      - name: id
        in: path
        description: The small ID of the node whose disputes should be retrieved
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: List of against attestation disputes where the specified node is the defendant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StackAttestationDispute'
        '500':
          description: Internal server error
  /attestation_disputes/own:
    get:
      tags:
      - attestation_disputes
      summary: Retrieves all attestation disputes initiated by the currently registered nodes.
      operationId: attestation_disputes_own_list
      responses:
        '200':
          description: List of all own attestation disputes where the registered nodes are the plaintiffs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StackAttestationDispute'
        '500':
          description: Internal server error
  /attestation_disputes/own/{id}:
    get:
      tags:
      - attestation_disputes
      summary: Retrieves all attestation disputes initiated by a specific node.
      operationId: attestation_disputes_own_get
      parameters:
      - name: id
        in: path
        description: The small ID of the node whose initiated disputes should be retrieved
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: List of own attestation disputes where the specified node is the plaintiff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StackAttestationDispute'
        '500':
          description: Internal server error
...
  /stacks/stacks/{id}:
    get:
      tags:
      - stacks
      summary: Retrieves all stacks for a specific node identified by its small ID.
      operationId: stacks_get
      parameters:
      - name: id
        in: path
        description: The small ID of the node whose stacks should be retrieved
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: List of Stack objects for the specified node
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stack'
        '500':
          description: Internal server error
...
components:
  schemas:
...
    Stack:
      type: object
      description: Represents a stack of compute units for a specific task
      required:
      - owner_address
      - stack_small_id
      - stack_id
      - task_small_id
      - selected_node_id
      - num_compute_units
      - price
      - already_computed_units
      - in_settle_period
      - total_hash
      - num_total_messages
      properties:
        already_computed_units:
          type: integer
          format: int64
          description: Number of compute units already processed
        in_settle_period:
          type: boolean
          description: Indicates whether the stack is currently in the settle period
        num_compute_units:
          type: integer
          format: int64
          description: Total number of compute units in this stack
        num_total_messages:
          type: integer
          format: int64
          description: Number of payload requests that were received by the node for this stack.
        owner_address:
          type: string
          description: Address of the owner of the stack
        price:
          type: integer
          format: int64
          description: Price of the stack (likely in smallest currency unit)
        selected_node_id:
          type: integer
          format: int64
          description: Identifier of the selected node for computation
        stack_id:
          type: string
          description: Unique string identifier for the stack
        stack_small_id:
          type: integer
          format: int64
          description: Unique small integer identifier for the stack
        task_small_id:
          type: integer
          format: int64
          description: Small integer identifier of the associated task
        total_hash:
          type: array
          items:
            type: integer
            format: int32
            minimum: 0
          description: |-
            Joint concatenation of Blake2b hashes of each payload and response pairs that was already processed
            by the node for this stack.
    StackAttestationDispute:
      type: object
      description: Represents a dispute in the stack attestation process
      required:
      - stack_small_id
      - attestation_commitment
      - attestation_node_id
      - original_node_id
      - original_commitment
      properties:
        attestation_commitment:
          type: array
          items:
            type: integer
            format: int32
            minimum: 0
          description: Cryptographic commitment provided by the attesting node
        attestation_node_id:
          type: integer
          format: int64
          description: Identifier of the node that provided the attestation
        original_commitment:
          type: array
          items:
            type: integer
            format: int32
            minimum: 0
          description: Original cryptographic commitment provided by the computing node
        original_node_id:
          type: integer
          format: int64
          description: Identifier of the original node that performed the computation
        stack_small_id:
          type: integer
          format: int64
          description: Unique small integer identifier for the stack involved in the dispute
...